+++ packages/maskbook/src/extension/service.ts
@@ -30,7 +30,6 @@ export const Services = {
     Helper: add(() => import('./background-script/HelperService'), 'Helper'),
     Provider: add(() => import('./background-script/ProviderService'), 'Provider'),
     Ethereum: add(() => import('./background-script/EthereumService'), 'Ethereum'),
-    Storage: add(() => import('./background-script/StorageService'), 'Storage'),
     SocialNetwork: add(() => import('./background-script/SocialNetworkService'), 'SocialNetwork'),
     Settings: add(() => import('./background-script/SettingsService'), 'Settings'),
     ThirdPartyPlugin: add(() => import('./background-script/ThirdPartyPlugin'), 'ThirdPartyPlugin'),
@@ -48,7 +47,6 @@ if (import.meta.webpackHot && isEnvironment(Environment.ManifestBackground)) {
             './background-script/HelperService',
             './background-script/ProviderService',
             './background-script/EthereumService',
-            './background-script/StorageService',
             './background-script/SettingsService',
             './background-script/ThirdPartyPlugin',
             './background-script/SocialNetworkService',
